+++ CHANGES.md
@@ -4,10 +4,11 @@
 
 ### Improvements
 
--   Add some missing properties to `read_info` (#365)
+-   Add some missing properties to `read_info`, including layer name, geometryname
+    and FID column name (#365).
 -   `read_arrow` and `open_arrow` now provide
     [GeoArrow-compliant extension metadata](https://geoarrow.org/extension-types.html),
-    including the CRS, when using GDAL 3.8 or higher.
+    including the CRS, when using GDAL 3.8 or higher (#366).
 
 ### Bug fixes